+++ CHANGELOG.md
@@ -4,5 +4,9 @@ Any notable changes to this project will be documented in this file.
 
 ## 1.0.0 - Initial Release (2024-07-25)
 
-- Initial release of Shadowpaw - Unity Toolkit
-- License changed to the [UNLICENSE](https://unlicense.org) (Public Domain)
+Initial release of Shadowpaw - Unity Toolkit
+
+- `License`: Changed to the [UNLICENSE](https://unlicense.org) (Public Domain)
+- `Utils`: Added extensions for LayerMask, List, Reflection, String, Task, and Type
+
+Be good to each other 💕